+++ indico/modules/categories/client/js/index.js
@@ -20,6 +20,7 @@ import SearchBox from 'indico/modules/search/components/SearchBox';
 import CategoryModeration from './components/CategoryModeration';
 import CategoryStatistics from './components/CategoryStatistics';
 import {LocaleContext} from './context.js';
+import './ind-category-navigator';
 
 (function(global) {
   document.addEventListener('DOMContentLoaded', () => {